Move bayerContourNextASTM.js to bayerContourNext.js Update unit tests to work with new device. Modifications to verifyChecksum to avoid the trim()
