+++ lib/core/device.js
@@ -34,7 +34,6 @@ var insuletOmniPod = require('../drivers/insuletDriver');
 var oneTouchUltra2 = require('../drivers/oneTouchUltra2');
 var abbottFreeStyleLite = require('../drivers/abbottFreeStyleLite');
 var bayerContourNext = require('../drivers/bayerContourNext');
-var bayerContourNextASTM = require('../drivers/bayerContourNextASTM');
 
 var device = {
   log: require('../bows')('Device')
@@ -53,7 +52,7 @@ device._deviceDrivers = {
   'BayerContourNextUsb': bayerContourNext,
   'BayerContourUsb': bayerContourNext,
   'BayerContourNextLink': bayerContourNext,
-  'BayerContourNextLink24': bayerContourNextASTM
+  'BayerContourNextLink24': bayerContourNext
 };
 
 device._deviceComms = {
