+++ lib/drivers/bayerContourNext.js
@@ -35,6 +35,7 @@ var debug = isBrowser ? require('../bows')('BCNextDriver') : debug;
 module.exports = function (config) {
   var cfg = _.clone(config);
   var hidDevice = config.deviceComms;
+  var frameBuffer = [];
   var HID_PACKET_SIZE = 64;
   // With no date & time settings changes available,
   // timezone is applied across-the-board
@@ -44,49 +45,97 @@ module.exports = function (config) {
    * TODO This info should be collected in the initial phase cause we can't assert if
    *      it is the same for all contour devices
    * */
-  var REPORT_BYTES = {
-    reportID: 0x41, // A
-    //checksum: 0x00, // Checksum is not been used by bayer contour device
-    hostID  : 0x42, // B
-    deviceID: 0x43  // C
-  };
+  var MAGIC_HEADER = 'ABC';
 
   /* end */
 
   var ASCII_CONTROL = {
     ACK : 0x06,
-    NAK : 0x15,
+    CR : 0x0D,
     ENQ : 0x05,
-    STX : 0x02,
+    EOT : 0x04,
     ETB : 0x17,
     ETX : 0x03,
-    EOT : 0x04,
-    TER : 0x4c
+    LF : 0x0A,
+    NAK : 0x15,
+    STX : 0x02
   };
+
+  var astmFrameBuffer = {
+    reset: function(){
+      this.bytes = new Uint8Array(0);
+      this.valid = false;
+      this.frame_len = 0;
+      this.payload = null;
+      return this;
+    },
+    setValid: function(){
+      this.payload = String.fromCharCode.apply(null, this.bytes);
+      this.valid = true;
+    },
+    clone: function(){
+      return _.clone(this);
+    }
+  }.reset();
 
   var probe = function(cb){
     debug('attempting probe of Bayer Contour Next');
   };
 
-  var bcnPacketHandler = function (buffer) {
-    var discardCount = _.keys(REPORT_BYTES).length;
+  // This is being used to reconstuct ASTM Frames
+  var bcnASTMFrameHandler = function (buffer) {
+    // Remove the MAGIC_HEADER from the front of the packet
+    var discardCount = MAGIC_HEADER.length;
     buffer.discard(discardCount);
 
-    if (buffer.len() === 0) { //empty buffer finish the data gathering
+    if (buffer.len() === 0) { // Empty buffer, finish the data gathering
       return false;
     }
 
-    var packet = extractPacket(buffer.bytes());
-    if (packet.packet_len !== 0) {
+    var astmFrame = extractPacketIntoFrame(buffer.bytes());
+    if (astmFrame.packet_len !== 0) {
       // cleanup the buffer data
       buffer.discard(HID_PACKET_SIZE - discardCount);
     }
 
-    if (packet.valid) {
-      return packet;
+    if (astmFrame.valid) {
+      return astmFrameBuffer.clone();
     } else {
       return null;
     }
+  };
+
+  var extractPacketIntoFrame = function (bytes) {
+    var packet_len = struct.extractByte(bytes, 0);
+    var byte1 = struct.extractByte(bytes, 1);
+    switch(byte1){
+      case ASCII_CONTROL.EOT:
+        astmFrameBuffer.reset();
+        astmFrameBuffer.setValid();
+        break;
+      case ASCII_CONTROL.STX:
+        astmFrameBuffer.reset();
+        break;
+    }
+
+    // Copy to the Frame Buffer, discabrding the length byte from the begining
+    var tmpbuff = new Uint8Array(astmFrameBuffer.frame_len + packet_len);
+    struct.copyBytes(tmpbuff, 0, astmFrameBuffer.bytes, astmFrameBuffer.frame_len, 0);
+    struct.copyBytes(tmpbuff, astmFrameBuffer.frame_len, bytes, packet_len, 1);
+    astmFrameBuffer.bytes = tmpbuff;
+    astmFrameBuffer.frame_len += packet_len;
+
+    // We're only using FRAME_TYPE for now, but we could use the other items for extra checks
+    var packetTail = struct.unpack(astmFrameBuffer.bytes, astmFrameBuffer.frame_len - 6, '2b2Z2Z', ['CR', 'FRAME_TYPE', 'CHECKSUM', 'CRLF']);
+
+    switch(packetTail['FRAME_TYPE']) {
+      case ASCII_CONTROL.ETX: // Last ASTM Frame
+      case ASCII_CONTROL.ETB: // End of valid ASTM Frame
+        astmFrameBuffer.setValid();
+        break;
+    }
+
+    return astmFrameBuffer;
   };
 
   var buildPacket = function (command, cmdlength) {
@@ -94,61 +143,18 @@ module.exports = function (config) {
                                  // if this value is changed the driver always returns the header
     var buf = new ArrayBuffer(datalen);
     var bytes = new Uint8Array(buf);
-    var ctr = struct.pack(bytes, 0, 'bbb', REPORT_BYTES.reportID,
-                          REPORT_BYTES.hostID, REPORT_BYTES.deviceID);
+    var ctr = struct.pack(bytes, 0, '3Z', MAGIC_HEADER );
     if (cmdlength) {
       ctr += struct.pack(bytes, ctr, 'bb', cmdlength, command);
     }
     return buf;
   };
 
-  var packetParser = function(result){
-    var tostr = _.map(result.bytes,
-                      function(e){
-                        return String.fromCharCode(e);
-                      }).join('');
-    result.payload = tostr;
-    return tostr;
-  };
-
   var buildAckPacket = function() {
     return buildPacket(ASCII_CONTROL.ACK, 1);
   };
 
-  var extractPacket = function (bytes) {
-    var packet = {
-      bytes: bytes,
-      valid: false,
-      packet_len: 0,
-      payload: null
-    };
-
-    var plen = bytes.length;
-    var packet_len = struct.extractByte(bytes, 0);
-
-    if (packet_len > plen) {
-      return packet;
-    }
-
-    //discard the length byte from the begining
-    var tmpbuff = new ArrayBuffer(packet_len);
-    struct.copyBytes(tmpbuff, 0, bytes, packet_len, 1);
-    packet.bytes = tmpbuff;
-    packet.packet_len = packet_len;
-    packet.valid = true;
-
-    return packet;
-  };
-
-
-  var buildHeaderCmd = function() {
-    return {
-      packet: buildAckPacket(),
-      parser: packetParser
-    };
-  };
-
-    //header data looks like
+  // header data looks like
   /*
   <STX>1H|\^&||qvqOi8|Bayer7350^01.14\01.03\04.18^7358-1611135^0000-
   |A=1^C=00^G=es,en\es\it\de\fr\hr\da\nl\fi\el\no\pt\sl\sv^I=0200^R=
@@ -156,19 +162,11 @@ module.exports = function (config) {
   |||||P|1|201505291248<ETB>01<CR><LF>
   */
 
-  var parseHeader = function (s){
-    var data = s.split('\n').filter(function(e){ return e !== '';});
-    var header = data.shift();
-
-
-
+  var parseHeader = function (header){
     if(verifyChecksum(header)){
-      var patient = data.shift();
-      var lineFeed = data.pop();
       var pString = header.split('|');
       var pInfo = pString[4].split('^');
       var sNum = pInfo[2].match(/^\d+\-\s*(\w+)/);
-      var records = data;
       var threshold = null;
       var thrs = pString[5].split('^');
 
@@ -184,30 +182,32 @@ module.exports = function (config) {
         model: pInfo[0],
         serialNumber: sNum[1],
         nrecs: pString[6],
-        rawrecords: records,
+        rawrecords: [],
       };
 
       if(threshold){
         devInfo.lowThreshold = parseInt(threshold[1]);
         devInfo.hiThreshold = parseInt(threshold[2]);
-      }else{
+      } else {
         devInfo.unreportedThreshold = true;
         devInfo.lowThreshold = 20;
         devInfo.hiThreshold = 600;
       }
 
       return devInfo;
-    }else{
+    } else {
       return null;
     }
   };
 
   function verifyChecksum(record){
-      var str = record.trim();
-      var data = str.split(String.fromCharCode(ASCII_CONTROL.ETB));
-      var check = data[1];
+      var buf = new Uint8Array(struct.packString(record));
+      var packetTail = struct.unpack(buf, record.length - 6, '2b2Z2Z', ['CR', 'ETB_OR_ETX', 'CHECKSUM', 'CRLF']);
+      var data = record.split(String.fromCharCode(packetTail['ETX_OR_ETB']));
+      var check = parseInt(packetTail['CHECKSUM'], 16);
+
       var sum = 0;
-      var n = record.slice(0, record.length - 3);
+      var n = record.slice(0, record.length - 4);
 
       _.map(n, function(e){
           if(e.charCodeAt(0) !== ASCII_CONTROL.STX){
@@ -215,9 +215,9 @@ module.exports = function (config) {
           }
       });
 
-      if((sum % 256) !== parseInt(check, 16)){
+      if((sum % 256) !== check){
           return null;
-      }else{
+      } else {
           return data[0];
       }
   }
@@ -228,12 +228,12 @@ module.exports = function (config) {
   var parseDataRecord = function (str, callback){
     var data = verifyChecksum(str);
     if(data){
-      var result = data.trim().match(/^.*\d+R\|(\d+).*Glucose\|(\d+)\|(\w+\/\w+).*\|{2}(.*)\|{2}(\d{12})$/).slice(1,6);
+      // TODO - The NextLink 2.4 also includes seconds in its timestamp (14 digits)
+      var result = data.match(/^[\x02]\dR\|(\d+)\|\^\^\^Glucose\|([0-9.]+)\|(\w+\/\w+).*\|{2}(.*)\|{2}(\d{12}).*/).slice(1,6);
       callback(null, result);
-    }else{
+    } else {
       throw( new Error('Invalid record data') );
     }
-
   };
 
   var getAnnotations = function (annotation, data){
@@ -271,110 +271,140 @@ module.exports = function (config) {
     if(markers.indexOf('C') !== -1) {
       debug('Marking as control test');
       return true;
-    }else{
+    } else {
       return false;
     }
   };
 
-  var getOneRecord = function (record, data, callback) {
-    parseDataRecord(record,function(err,r){
-      var robj = {};
-      if (err) {
-          debug('Failure trying to read record', record);
-          debug(err);
-          return callback(err, null);
+  var getOneRecord = function (data, callback) {
+    var cmd = buildAckPacket();
+    bcnCommandResponse(cmd, function (err, record) {
+      // TODO - add error handling
+      // Skip over the STX and the sequence number.
+      var record_type = struct.extractByte(record, 2);
+
+      if(record_type === 'R'){
+        parseDataRecord(record,function(err,r){
+          var robj = {};
+          if (err) {
+            debug('Failure trying to read record', record);
+            debug(err);
+            return callback(err, null);
+          } else {
+            robj.timestamp = parseInt(r[4]);
+            robj.annotations = getAnnotations(r[3], data);
+            robj.control = isControl(r[3]);
+            robj.units = r[2];
+            if(robj.units === 'mmol/L') {
+              robj.glucose = parseFloat(r[1]);
+            } else {
+              robj.glucose = parseInt(r[1]);
+            }
+            robj.nrec = parseInt(r[0]);
+            return callback(null, robj);
+          }
+        });
       } else {
-          robj.timestamp = parseInt(r[4]);
-          robj.annotations = getAnnotations(r[3], data);
-          robj.control = isControl(r[3]);
-          robj.units = r[2];
-          robj.glucose = parseInt(r[1]);
-          robj.nrec = parseInt(r[0]);
-          return callback(null, robj);
+        return callback(null, null);
       }
     });
   };
 
+  var debugFrame = function (astmFrame) {
+    var output = astmFrame
+      .replace(/[\x02]/g, '<STX>')
+      .replace(/[\x03]/g, '<ETX>')
+      .replace(/[\x04]/g, '<EOT>')
+      .replace(/[\x05]/g, '<ENQ>')
+      .replace(/[\x06]/g, '<ACK>')
+      .replace(/[\x0A]/g, '<LF>')
+      .replace(/[\x0D]/g, '<CR>')
+      .replace(/[\x15]/g, '<NAK>')
+      .replace(/[\x17]/g, '<ETB>');
+    return output;
+  };
+
   var bcnCommandResponse = function (commandpacket, callback) {
-
-    hidDevice.send(commandpacket.packet, function () {
-        receivePacket(5000, commandpacket.parser, function(err, packet) {
-            var rawtext = '';
-            while(hidDevice.hasAvailablePacket()){
-                var pkt = hidDevice.nextPacket();
-                pkt.parsed_payload = commandpacket.parser(pkt);
-                rawtext += pkt.parsed_payload;
-            }
-
-            if (err !== 'TIMEOUT') {
-              callback(err, rawtext);
-            }else{
-              debug('receivePacket timed out');
-              callback(err, null);
-            }
-        });
+    hidDevice.send(commandpacket, function () {
+      getASTMFrame(5000, 3, function(err) {
+        if (err !== null || !hidDevice.hasAvailablePacket()) {
+          // TODO - we should do retries here, depending on the error
+          callback(err, null);
+        } else {
+          // We're using the hidDevice as a message/frame buffer, not a USB packet buffer
+          // to save on buffer copying. It could be refactored differently easily enough.
+          var frame = hidDevice.nextPacket();
+          // TODO - we should ideally check the validity of the ASTM frame (sequence number, checksum etc) here as well for retries
+          //debug(debugFrame(frame.payload));
+          callback(err, frame.payload);
+        }
+      });
     });
   };
 
-  var receivePacket = function (timeout, parser, callback) {
+  var getASTMFrame = function (timeout, retries, callback) {
     var abortTimer = setTimeout(function () {
       clearInterval(listenTimer);
       debug('TIMEOUT');
-      callback('TIMEOUT', null);
+      callback('TIMEOUT');
     }, timeout);
 
     var listenTimer = setInterval(function () {
       hidDevice.receive(function(raw) {
-          var r = new Uint8Array(raw);
-          var startIndex = 0;
-          var rec = struct.unpack(r, 6, 'b', ['TYPE']);
-          var hasETB = _.indexOf(r, ASCII_CONTROL.ETB, startIndex);
-          if (hasETB !== -1) {
-            clearTimeout(abortTimer);
-          }
-          debug(rec);
-          if (hasETB !== -1) {
-            // send a new ACK
-            var cmd = buildAckPacket();
-            hidDevice.send(cmd, function () {
-                // we need to re-set timeout for each packet, as the number of packets are of variable length
-                abortTimer = setTimeout(function () {
-                  clearInterval(listenTimer);
-                  debug('TIMEOUT');
-                  callback('TIMEOUT', null);
-                }, timeout);
-                debug('New ACK SENT');
-            });
-          }
-          else if(rec.TYPE == ASCII_CONTROL.TER){
+        var packet = new Uint8Array(raw);
+
+        var packetHead = struct.unpack(packet, 0, '3Z2b', ['HEADER', 'SIZE', 'BYTE1']);
+
+        if(packetHead['HEADER'] != MAGIC_HEADER){
+          debug('Invalid packet from Contour device');
+          callback('INVALID_USB_PACKET');
+        }
+
+        switch(packetHead['BYTE1']) {
+          case ASCII_CONTROL.EOT:
+          case ASCII_CONTROL.ENQ:
             clearTimeout(abortTimer);
             clearInterval(listenTimer);
-            callback(null, '');
-          }
-        });
-      }, 20); //we need to be ready for the next packet quickly
+            callback(null);
+            break;
+        }
+
+        // The tail of the packet starts 6 from the end, but because we haven't stripped the
+        // MAGIC_HEADER and length byte from packet, we're using SIZE - 2
+        var packetTail = struct.unpack(packet, parseInt(packetHead['SIZE']) - 2, '2b2Z2Z', ['CR', 'FRAME_TYPE', 'CHECKSUM', 'CRLF']);
+        switch(packetTail['FRAME_TYPE']) {
+          case ASCII_CONTROL.ETX: // Last ASTM Frame
+          case ASCII_CONTROL.ETB: // End of valid ASTM Frame
+            clearTimeout(abortTimer);
+            clearInterval(listenTimer);
+            callback(null);
+            break;
+        }
+      });
+    }, 20); // The Contour timeout is 15 seconds, but let's be slow enough to be kind to the CPU.
   };
 
   var getDeviceInfo = function (obj, cb) {
-      debug('DEBUG: on getDeviceInfo');
-      var cmd = buildHeaderCmd();
-      bcnCommandResponse(cmd, function (err, datatxt) {
-          if (err) {
-              debug('Failure trying to talk to device.');
-              debug(err);
-              debug(datatxt);
-              cb(null, null);
-          } else {
-              obj.header = datatxt;
-              if(parseHeader(datatxt)){
-                _.assign(obj, parseHeader(datatxt));
-                cb(null, obj);
-              }else{
-                debug('Invalid header data');
-                throw(new Error('Invalid header data'));
-              }
-          }
-      });
+    debug('DEBUG: in getDeviceInfo');
+    var cmd = buildAckPacket();
+    bcnCommandResponse(cmd, function (err, datatxt) {
+      if (err) {
+        debug('Failure trying to talk to device.');
+        debug(err);
+        debug(datatxt);
+        cb(null, null);
+      } else {
+        obj.header = datatxt;
+        var header = parseHeader(datatxt);
+        if(header){
+          _.assign(obj, header);
+          cb(null, obj);
+        } else {
+          debug('Invalid header data');
+          throw(new Error('Invalid header data'));
+        }
+      }
+    });
   };
 
   var processReadings = function(readings) {
@@ -416,7 +446,7 @@ module.exports = function (config) {
           }
         dataToPost.push(smbg);
       }
-    }else{
+    } else {
       debug('Device has no records to upload');
       throw(new Error('Device has no records to upload'));
     }
@@ -439,14 +469,13 @@ module.exports = function (config) {
     connect: function (progress, data, cb) {
       debug('in connect!');
 
-      cfg.deviceComms.connect(data.deviceInfo, bcnPacketHandler, probe, function(err) {
+      cfg.deviceComms.connect(data.deviceInfo, bcnASTMFrameHandler, probe, function(err) {
         if (err) {
           return cb(err);
         }
         data.disconnect = false;
         progress(100);
         cb(null, data);
-
       });
     },
 
@@ -461,7 +490,7 @@ module.exports = function (config) {
               _.assign(data, result);
 
               cb(null, data);
-          }else{
+          } else {
               cb(err,result);
           }
       });
@@ -471,19 +500,21 @@ module.exports = function (config) {
       debug('in fetchData', data);
 
       function getOneRecordWithProgress(recnum, cb) {
-        var rec = data.rawrecords.shift();
         progress(100.0 * recnum / data.nrecs);
         setTimeout(function() {
-          getOneRecord(rec, data, cb);
+          getOneRecord(data, cb);
         }, 20);
       }
 
-      async.timesSeries(data.nrecs, getOneRecordWithProgress, function(err, result) {
+      // FIXME - We're adding +1 to skip the Patient Record (P)
+      async.timesSeries(parseInt(data.nrecs)+1, getOneRecordWithProgress, function(err, result) {
         if (err) {
           debug('fetchData failed');
           debug(err);
           debug(result);
         } else {
+          // FIXME - delete the null record at the front for the patient record.
+          result.splice(0,1);
           debug('fetchData', result);
         }
         data.fetchData = true;
@@ -494,6 +525,7 @@ module.exports = function (config) {
     },
 
     processData: function (progress, data, cb) {
+        //debug('in processData');
         progress(0);
         data.bg_data = processReadings(data.bgmReadings);
         data.post_records = prepBGData(progress, data);
@@ -558,7 +590,7 @@ module.exports = function (config) {
               data.disconnect = true;
               cb(null, data);
           });
-      }else{
+      } else {
         progress(100);
       }
     }
